Add base class for lambdas
This commit implements LambdaBase to deduplicate code common between
lambdas and removes unnecessary directory nesting in the exodus_lambda
directory.
