+++ exodus_lambda/__init__.py
@@ -1,8 +1,8 @@
-from exodus_lambda.functions.origin_request.origin_request import (
+from exodus_lambda.functions.origin_request import (
     lambda_handler as origin_request,
 )
 
-from exodus_lambda.functions.origin_response.origin_response import (
+from exodus_lambda.functions.origin_response import (
     lambda_handler as origin_response,
 )
