+++ exodus_lambda/functions/origin_response.py
@@ -1,23 +1,12 @@
-import json
 import logging
 import re
+
+from .base import LambdaBase
 
 LOG = logging.getLogger("origin-response")
 
 
-class LambdaClient(object):
-    def __init__(self, conf_file="lambda_config.json"):
-        self._conf_file = conf_file
-        self._conf = None
-
-    @property
-    def conf(self):
-        if not self._conf:
-            with open(self._conf_file, "r") as json_file:
-                self._conf = json.load(json_file)
-
-        return self._conf
-
+class OriginResponse(LambdaBase):
     def handler(self, event, context):
         # pylint: disable=unused-argument
 
@@ -61,4 +50,4 @@ class LambdaClient(object):
 
 
 # Make handler available at module level
-lambda_handler = LambdaClient().handler  # pylint: disable=invalid-name
+lambda_handler = OriginResponse().handler  # pylint: disable=invalid-name
